requirements.txt
tensorflow==2.5.0
numpy==1.19.5
matplotlib==3.4.2
pandas==1.2.4
tqdm==4.61.0
Pillow==8.2.0
